tscache: create Cache interface which cacheImpl implements
This change pulls `tscache.Cache` into an interface. This is helpful for
two reasons. First, it separates the interface from the implementation,
which improves readability of the cache and makes its external facing
methods more obvious. This also allows future implementations of `Cache`
(such as one written in C++) to be swapped in-place of the current
implementation without any changes outside of this package.
